Convert App and LabelButton to functional components
- Converted App from class to functional component (50→66 lines)
  - Replaced constructor + history.listen with useEffect
  - Properly cleanup history listener on unmount
  - Migrated from .jsx to .tsx with proper TypeScript types
  - Added Alert and ReduxState interfaces

- Converted LabelButton from class to functional component (26→24 lines)
  - Removed unnecessary render method
  - Replaced Function type with proper () => void
  - Added LabelButtonProps interface

All class components in src/_components are now functional components
Build succeeds with no warnings
Reduced bundle size by 176 bytes total (32 + 144 bytes)
